Refactor indexing script: moves logic to testable module
No functional changes.
